Fix issue with finding the real python executable during tests
Some platforms do not provide `/usr/bin/python` and instead provide `/usr/bin/python3`.
When running under nox, which creates a virtualenv to run the tests, we try to find the real python executable, not the virtualenv one used to run the test suite, when creating new virtualenv's.
This fixes that issue.